Implemented robust sanitization methods throughout
The server now properly sanitizes all input data during POST requests.
Even though this code is designed to be run on secure LAN networks,
header spoofing is accounted for. Node sanitization packages and
regular expressions are used to strip bad characters that could damage
the internal file system of the content folder, or be used to gain
access to files outside of the scope of the content.